mapp.location.create() method and decorator docs.
