+++ lib/layer/decorate.mjs
@@ -65,66 +65,6 @@ export default async function decorate(layer) {
     console.warn(`Layer: ${layer.key}, please move draw.delete to use layer.deleteLocation:true.`);
   }
 
-  // Callback which creates and stores a location from a feature returned by the drawing interaction.
-  layer.draw.callback = async (feature, params) => {
-    // If the feature is null, return.
-    if (!feature) return;
-
-    // Get the current table and set new to true.
-    const location = {
-      layer,
-      table: layer.tableCurrent(),
-      new: true
-    };
-
-    // Store location in database.
-    // The id must be returned from a serial ID field.
-    location.id = await mapp.utils.xhr({
-      method: 'POST',
-      url: `${layer.mapview.host}/api/query?` +
-        mapp.utils.paramString({
-          template: 'location_new',
-          locale: layer.mapview.locale.key,
-          layer: layer.key,
-          table: location.table
-        }),
-      body: JSON.stringify({
-        [layer.geom]: feature.geometry,
-
-        // Spread in defaults.
-        ...params?.defaults,
-        ...layer.draw?.defaults
-      })
-    });
-
-    // Check whether feature is loaded on MVT update.
-    if (layer.format === 'mvt') {
-      layer.features = [];
-      layer.source.on('tileloadend', concatFeatures);
-
-      function concatFeatures(e) {
-        layer.features = layer.features.concat(e.tile.getFeatures());
-      }
-
-      setTimeout(checkFeature, 1000);
-
-      function checkFeature() {
-        let found = layer.features?.find(F => F.properties?.id === location.id);
-        if (found) {
-          layer.source.un('tileloadend', concatFeatures);
-        } else {
-          layer.reload();
-        }
-      }
-    }
-
-    // Layer must be reloaded to reflect geometry changes.
-    layer.reload();
-
-    // Get the newly created location.
-    mapp.location.get(location);
-  };
-
   // Set layer filter.
   layer.filter = {
     current: {},
