+++ lib/location/_location.mjs
@@ -1,9 +1,10 @@
 /**
-## mapp.location{}
-The mapp.location module provides methods to get and decorate location objects.
+## /location
 
 @module /location
 */
+
+import create from './create.mjs'
 
 import decorate from './decorate.mjs'
 
@@ -12,6 +13,7 @@ import {get, getInfoj} from './get.mjs'
 import nnearest from './nnearest.mjs'
 
 export default {
+  create,
   decorate,
   get,
   getInfoj,
