+++ lib/location/decorate.mjs
@@ -1,11 +1,21 @@
 /**
-## mapp.location.decorate()
-@module /location/decorate
+### /location/decorate
 
-@param {Object} location
-The location object to be decorated.
+@module /location/decorate
 */
-export default location => {
+
+/**
+@function decorate
+
+@description
+The location decorator method creates mapp-location typedef object from a JSON location.
+
+@param {object} location JSON location.
+
+@returns {location} Decorated Mapp Location.
+*/
+
+export default function decorate(location) {
 
   Object.assign(location, {
     flyTo,
