+++ lib/ui/elements/drawing.mjs
@@ -1,3 +1,9 @@
+/**
+### /ui/elements/drawing
+
+@module /ui/elements/drawing
+*/
+
 export default {
   point,
   line,
@@ -161,6 +167,23 @@ mapp.utils.merge(mapp.dictionaries, {
   }
 })
 
+/**
+@function drawOnclick
+
+@description
+The drawOnClick method is triggered by clicking on a drawing element button.
+
+The 'active' class is toggled on the button element. The drawing interaction is finished if the 'active' class is toggled off.
+
+A callback method is assigned to the interaction before the interaction object is passed as argument to call the mapview's draw interaction.
+
+@param {Object} e The click event. 
+@param {layer} layer Decorated Mapp Layer.
+@param {Object} interaction Mapview drawing interaction.
+
+@property {Object} e.target The click event target [button].
+*/
+
 function drawOnclick(e, layer, interaction) {
 
   const btn = e.target
@@ -175,7 +198,7 @@ function drawOnclick(e, layer, interaction) {
 
   interaction.callback = feature => {
 
-    layer.draw.callback(feature, interaction)
+    mapp.location.create(feature, interaction, layer)
 
     btn.classList.remove('active')
